Updated ContentView Dependency Injection
